% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{priors}
\alias{priors}
\alias{normal}
\alias{student_t}
\alias{cauchy}
\alias{exponential}
\alias{log_normal}
\title{Prior distributions and options}
\usage{
normal(location = 0, scale = NULL, autoscale = TRUE)

student_t(df = 1, location = 0, scale = NULL, autoscale = TRUE)

cauchy(location = 0, scale = NULL, autoscale = TRUE)

exponential(rate = 1, autoscale = TRUE)

log_normal(location = 0, scale = 1)
}
\arguments{
\item{location}{Prior location. In most cases, this is the prior mean, but
for \code{cauchy} (which is equivalent to \code{student_t} with
\code{df=1}), the mean does not exist and \code{location} is the prior
median. The default value is \eqn{0}.}

\item{scale}{Prior scale. The default depends on the family (see
\strong{Details}).}

\item{autoscale}{A logical scalar, defaulting to \code{TRUE}. If \code{TRUE} 
then the scales of the priors on the intercept and regression coefficients 
may be additionally modified internally by \pkg{rstanarm} in the following 
cases. First, for Gaussian models only, the prior scales for the intercept, 
coefficients, and the auxiliary parameter \code{sigma} (error standard 
deviation) are multiplied by \code{sd(y)}. Additionally --- not only for 
Gaussian models --- if the \code{QR} argument to the model fitting function
(e.g. \code{stap_glm}) is \code{FALSE} then: for a predictor with only one 
value nothing is changed; for a predictor \code{x} with exactly two unique 
values, we take the user-specified (or default) scale(s) for the selected 
priors and divide by the range of \code{x}; for a predictor \code{x} with 
more than two unique values, we divide the prior scale(s) by \code{sd(x)}.}

\item{df}{Prior degrees of freedom. The default is \eqn{1} for 
\code{student_t}, in which case it is equivalent to \code{cauchy}.
For the \code{product_normal}   prior, the degrees of freedom 
parameter must be an integer (vector) that is  at least \eqn{2} (the default).}

\item{rate}{Prior rate for the exponential distribution. Defaults to 
\code{1}. For the exponential distribution the rate parameter is the 
\emph{reciprocal} of the mean.}
}
\value{
A named list to be used internally by the \pkg{rstap2} model
  fitting functions.
}
\description{
The functions described on this page are used to specify the prior-related #' arguments of the various modeling functions in the \pkg{rstap} package (to
view the priors used for an existing model see \code{\link{prior_summary}}). 
The default priors used in the various \pkg{rstap} modeling functions are
intended to be \emph{weakly informative} in that they provide moderate
regularlization and help stabilize computation. For many applications the
defaults will perform well, but prudent use of more informative priors is
encouraged. All of the priors here are informed by the priors in \pkg{rstanarm}, though it should be noted that the heirarchical shape priors are not included.
}
\references{
Gelman, A., Jakulin, A., Pittau, M. G., and Su, Y. (2008). A weakly
informative default prior distribution for logistic and other regression
models. \emph{Annals of Applied Statistics}. 2(4), 1360--1383.

# Can assign priors to names
N05 <- normal(0, 5)
}
\seealso{
The various vignettes for the \pkg{rstanarm} and \pkg{rstap2} packages also discuss 
  and demonstrate the use of some of the supported prior distributions.

#'
}
