% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stapdnd_glm2.fit.R
\name{stapdnd_glm2.fit}
\alias{stapdnd_glm2.fit}
\title{Fitting Difference in Difference STAP glms}
\usage{
stapdnd_glm2.fit(y, z, dists_crs, u_s, subj_matrix, subj_n,
  family = gaussian(), stap_par_code, iter_max, warmup, adapt_delta,
  max_treedepth, seed, diagnostics, cores = 1, prior_dnd = normal(),
  prior_intercept = normal(), prior_bar = normal(),
  prior_theta = lognormal(), prior_sigma = cauchy())
}
\arguments{
\item{y}{an N-length vector}

\item{z}{N x p design matrix of subject specific covariates}

\item{dists_crs}{(q_s+q_st) x M matrix of distances between outcome 
observations and built environment features with a hypothesized spatial scale}

\item{u_s}{N x (q *2) matrix of compressed row storage array indices for dists_crs}

\item{subj_matrix}{an n x N matrix indexing the subjects observations in the y and Z matrices}

\item{family}{distributional family - only binomial gaussian or poisson currently allowed}

\item{prior_dnd, prior_intercept, prior_bar, prior_theta, prior_sigma}{}

\item{stap_data}{object of class "stap_data" that contains information on all the spatial-temporal predictors in the model}

\item{max_distance}{the upper bound on any and all distances included in the model}

\item{max_time}{the upper bound on any and all times included in the model}

\item{weights}{weights to be added to the likelihood observation for a given subject}

\item{offset}{offset term to be added to the outcome for a given subject}
}
\description{
Fitting Difference in Difference STAP glms
}
