% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stapdnd_glmer2.fit.R
\name{stapdnd_glmer2.fit}
\alias{stapdnd_glmer2.fit}
\title{Bayesian spatial-temporal generalized linear models with group-specific terms via Stan}
\usage{
stapdnd_glmer2.fit(y, z, w, dists_crs, u_s, subj_matrix, subj_n,
  family = gaussian(), stap_par_code, iter_max, warmup, adapt_delta,
  max_treedepth, seed, diagnostics, cores = 1, prior_dnd = normal(),
  prior_intercept = normal(), prior_bar = normal(),
  prior_theta = lognormal(), prior_sigma = cauchy())
}
\arguments{
\item{y}{an N-length vector}

\item{z}{N x p design matrix of subject specific covariates}

\item{dists_crs}{(q_s+q_st) x M matrix of distances between outcome 
observations and built environment features with a hypothesized spatial scale}

\item{u_s}{N x (q *2) matrix of compressed row storage array indices for dists_crs}

\item{subj_matrix}{an n x N matrix indexing the subjects observations in the y and Z matrices}

\item{subj_n}{an n x q matrix where each row holds the reciprocal of the number of observations for the ith subject}

\item{family}{distributional family - only binomial gaussian or poisson currently allowed}

\item{stap_data}{object of class "stap_data" that contains information on all the spatial-temporal predictors in the model}

\item{max_distance}{the upper bound on any and all distances included in the model}

\item{max_time}{the upper bound on any and all times included in the model}

\item{weights}{weights to be added to the likelihood observation for a given subject}

\item{offset}{offset term to be added to the outcome for a given subject}
}
\description{
Bayesian inference for stap-glms with group-specific coefficients that have 
unknown covariance matrices with flexible priors.
}
\seealso{
The Longituinal \href{https://biostatistics4socialimpact.github.io/rstap/articles/longitudinal-I.html}{Vignette} for \code{stap_glmer} 
and the \href{http://arxiv.org/abs/1812.10208}{preprint} article available through arXiv.
}
